app/backend/requirements.txt
-i https://pypi.org/simple
aiofiles==23.2.1 ; python_version >= '3.7'
aiohttp==3.9.1
aiosignal==1.3.1 ; python_version >= '3.7'
annotated-types==0.6.0 ; python_version >= '3.8'
anyio==3.7.1 ; python_version >= '3.7'
attrs==23.1.0 ; python_version >= '3.7'
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.29.5
azure-identity==1.15.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0 ; python_version >= '3.7'
azure-mgmt-storage==21.1.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-storage-queue==12.8.0
blinker==1.7.0 ; python_version >= '3.8'
certifi==2023.11.17 ; python_version >= '3.6'
cffi==1.16.0 ; python_version >= '3.8'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
cryptography==41.0.7 ; python_version >= '3.7'
distro==1.8.0 ; python_version >= '3.6'
flask==3.0.0 ; python_version >= '3.8'
frozenlist==1.4.0 ; python_version >= '3.8'
gunicorn==21.2.0
h11==0.14.0 ; python_version >= '3.7'
h2==4.1.0 ; python_full_version >= '3.6.1'
hpack==4.0.0 ; python_full_version >= '3.6.1'
httpcore==1.0.2 ; python_version >= '3.8'
httpx==0.25.2 ; python_version >= '3.8'
hypercorn==0.15.0 ; python_version >= '3.7'
hyperframe==6.0.1 ; python_full_version >= '3.6.1'
idna==3.6 ; python_version >= '3.5'
isodate==0.6.1
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
msal==1.25.0 ; python_version >= '2.7'
msal-extensions==1.0.0
msrest==0.7.1 ; python_version >= '3.6'
multidict==6.0.4 ; python_version >= '3.7'
oauthlib==3.2.2 ; python_version >= '3.6'
openai==1.3.6
packaging==23.2 ; python_version >= '3.7'
portalocker==2.8.2 ; python_version >= '3.5' and platform_system != 'Windows'
priority==2.0.0 ; python_full_version >= '3.6.1'
pycparser==2.21
pydantic==2.5.2 ; python_version >= '3.7'
pydantic-core==2.14.5 ; python_version >= '3.7'
pyjwt[crypto]==2.8.0 ; python_version >= '3.7'
pymupdf==1.23.6
pymupdfb==1.23.6 ; python_version >= '3.8'
quart==0.19.4
quart-cors==0.7.0
regex==2023.10.3 ; python_version >= '3.7'
requests==2.31.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1 ; python_version >= '3.7'
typing-extensions==4.8.0 ; python_version >= '3.8'
urllib3==2.1.0 ; python_version >= '3.8'
uvicorn==0.24.0.post1
werkzeug==3.0.1 ; python_version >= '3.8'
wsproto==1.2.0 ; python_full_version >= '3.7.0'
yarl==1.9.3 ; python_version >= '3.7'
